backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==23.0.0
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
statsmodels==0.14.4
chardet==5.2.0
pyarrow==20.0.0
openpyxl==3.1.5
Werkzeug==3.1.3
gunicorn==23.0.0